Changes suggested by Samson and Mark:
Mark suggests that we rename the tab as "External Reviews"  to make the meaning more precise. Export and Import tab probably should go, as it's not usable with ICTM at this time.

In the ICTM Content Tab, Temporal, Severity, and Functional properties probably should go, as they are not mentioned in the ICTM content model, and their applicability to TM is questionable.

Renaming the "Diagnostic Criteria" to "Diagnostic Method" would more accurately the character of the terms that Molly placed under "Diagnositic Criteria" in the CM.